requirements.txt
antiorm==1.2.1
asgiref==3.2.3
codegen==1.0
db==0.1.1
Django==3.0.2
django-composite-foreignkey==1.1.0
django-compositekey==1.5.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
gis==0.2.1
gunicorn==20.0.4
numpy==1.18.1
patterns==0.3
psycopg2==2.8.4
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.0
